move the `result_type` compat code to `array_api_compat`
